requirements.txt
flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
PyPDF2==3.0.1
markdown==3.5.2
beautifulsoup4==4.12.2 